requirements.txt
django==1.9.2
djangorestframework==3.3.2
django-rest-auth==0.7.0